moved ferch()code in donordetails /components/ index.js
